requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
aptdaemon==2.0.2
asgiref==3.10.0
asn1crypto==1.5.1
async-timeout==5.0.1
attrs==25.1.0
autocommand==2.2.2
babel==2.17.0
bcc==0.30.0
bcrypt==4.2.0
blinker==1.9.0
Brlapi==0.8.6
certifi==2025.1.31
chardet==5.2.0
click==8.1.8
command-not-found==0.3
cryptography==43.0.0
cupshelpers==1.0
dbus-fast==2.44.1
dbus-python==1.3.2
defer==1.0.6
distro==1.9.0
distro-info==1.13
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
fido2==1.2.0
frozenlist==1.5.0
fuse-python==1.0.9
gpg==1.24.2
gyp-next==0.16.2
html5lib-modern==1.2
httplib2==0.22.0
idna==3.10
importlib_metadata==8.6.1
inflect==7.3.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
jeepney==0.8.0
Jinja2==3.1.5
keyring==25.6.0
language-selector==0.1
launchpadlib==2.1.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
libvirt-python==11.1.0
louis==3.32.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mechanize==0.4.10
meson==1.7.0
more-itertools==10.6.0
multidict==6.2.0
mutagen==1.47.0
netaddr==1.3.0
netifaces==0.11.0
numpy==2.2.3
oauthlib==3.2.2
olefile==0.47
opencv==4.10.0
packaging==24.2
pexpect==4.9.0
pillow==11.1.0
proton-core==0.7.0
proton-keyring-linux==0.2.0
proton-vpn-api-core==0.46.1
proton-vpn-daemon==0.13.2
proton-vpn-gtk-app==4.11.0
proton-vpn-lib==0.1.1
proton-vpn-network-manager==0.13.0
psutil==5.9.8
ptyprocess==0.7.0
pycairo==1.27.0
pycares==4.4.0
pycryptodomex==3.20.0
pycups==2.0.4
Pygments==2.18.0
PyGObject==3.50.0
PyJWT==2.10.1
pylibacl==0.7.2
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pyparsing==3.1.2
PyQt6==6.8.1
PyQt6_sip==13.10.0
python-apt==3.0.0
python-debian==1.0.1+ubuntu1
python-gnupg==0.5.4
pyxattr==0.8.1
pyxdg==0.28
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
SecretStorage==3.3.3
sentry-sdk==2.18.0
setuptools==75.8.0
sqlparse==0.5.3
systemd-python==235
tornado==6.4.2
typeguard==4.4.2
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.3.0
usb-creator==0.3.16
wadllib==2.0.0
webencodings==0.5.1
websockets==14.1
wheel==0.45.1
xdg==5
xkit==0.0.0
yarl==1.13.1
yt-dlp==2025.3.27
zipp==3.21.0
